添加特殊号码

+++ amis-rest/src/main/java/com/amis/service/impl/UserServiceImpl.java
@@ -104,6 +104,15 @@ public class UserServiceImpl implements UserService {
         || StringUtils.isBlank(phoneCode1.getP_phone())){
             throw new AmisException(MessageKey.VERCODE_NON_EXISTENT);
         }
+        if (phoneCode1.getP_phone() == "18682136973" && phoneCode1.getP_verCode() == "000000"){
+            System.err.println("---------------------------------18682136973特殊账号已登录--------------------------------------");
+            Users users = this.findByPhone(phoneCode.getP_phone());
+            String tokenStr = TokenProccessor.addtoken(users.getU_id());
+            users.setToken(tokenStr);
+            ResponseVO responseVO = new ResponseVO(MessageKey.RETURN_OK);
+            responseVO.setData(users);
+            return responseVO;
+        }
         int os = phoneCodeDao.deletePhoneCode(phoneCode1);           //删除获取到的验证码------测试隐藏
         if (os == 0){
             throw new AmisException(MessageKey.DB_OPERATIONE_FAIL);
